chat-service/requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
pydantic==2.12.4
python-socketio==5.12.0
python-dotenv==1.1.1
httpx==0.28.1
openai==2.7.2
